requirement.txt
Django==1.8.4
django-tastypie==0.12.2
requests==2.9.1
tweepy==3.5.0
librabbitmq==1.6.1
eventlet==0.19.0